Specify asset host by STATIC_ASSET_ROOT env var
As part of the work to enable RFC 115 [1] we intend to switch static
from serving assets on the asset hostname to instead use the regular www
hostname. Static, unlike other GOV.UK applications, requires a host name
to be set so that the live version of static can be used in development
environments and heroku preview environments.

I felt that it no longer made sense to use the GOVUK_ASSET_ROOT
environment variable since this is no longer a GOV.UK wide environment
variable so I set the name to be static specific.

Both environment variables are in use in these changes so that static
can continue to operate under both configurations.

[1]: https://github.com/alphagov/govuk-rfcs/blob/45e09c088bfde6c1d4dfb68f9bc1c635f63c9a39/rfc-115-enabling-http2-on-govuk.md